fucking t -> T
