+++ Bilbary for iPhone/Viiews/WeekView.swift
@@ -60,7 +60,4 @@ struct WeekView: View {
     }
 }
 
-#Preview {
-    
-    WeekView(isTodayActive: true)
-}
+
