+++ Bilbary for iPhone/iPhone Models/determineHigherValue.swift
